requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --upgrade-package Jinja2
#
boto3==1.9.169
botocore==1.12.169        # via boto3, s3transfer
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via rq
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
dj-database-url==0.5.0    # via django-heroku
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-storages==1.7.1
django==2.2.2
djangorestframework-gis==0.14
djangorestframework==3.9.2
docutils==0.14            # via botocore
faker==1.0.4
gunicorn==19.9.0
idna==2.8                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10.1
jmespath==0.9.4           # via boto3, botocore
markdown==3.0.1
markupsafe==1.1.1         # via jinja2
numpy==1.16.2
openapi-codec==1.3.2      # via django-rest-swagger
psycopg2-binary==2.8.3
psycopg2==2.8.3           # via django-heroku
python-dateutil==2.8.0    # via botocore, faker
pytz==2018.9              # via django
redis==3.2.1              # via rq
requests==2.21.0          # via coreapi
rq==1.0
s3transfer==0.2.1         # via boto3
simplejson==3.16.0        # via django-rest-swagger
six==1.12.0               # via faker, python-dateutil
sqlparse==0.3.0           # via django
text-unidecode==1.2       # via faker
uritemplate==3.0.0        # via coreapi
urllib3==1.24.3           # via botocore, requests
whitenoise==4.1.2
